+++ src/browserProvider.ts
@@ -3,15 +3,45 @@ import * as vscode from "vscode";
 
 export class BrowserProvider {
     private context: vscode.ExtensionContext;
-    private browserContext: BrowserContext;
+    private browserContext: BrowserContext | undefined;
     private loggedIn: boolean = false;
 
     constructor(
         context: vscode.ExtensionContext,
-        browserContext: BrowserContext
     ) {
         this.context = context;
-        this.browserContext = browserContext;
+    }
+
+    async initializeBrowser() {
+        if (!this.browserContext) {
+            let browser: Browser;
+            try {
+                browser = await chromium.launch({ headless: true });
+            } catch (e) {
+                await this.installHelper(e);
+                return;
+            }
+            const browserContext = await browser.newContext();
+            this.browserContext = browserContext;
+        }
+    }
+
+    async installHelper(e: any) {
+        vscode.window.showErrorMessage("Playwright is probably not installed.\n" + e);
+        const result: string = (await vscode.window.showQuickPick([
+            "Yes, please install Playwright browsers for WebLab (requires NPM)",
+            "No (This will deactivate WebLab plugin)"],
+            {
+                title: "Playwright browsers are probably not installed. Would you like to install?"
+            }) ?? "No");
+        if (result.startsWith("Yes")) {
+            const terminal = vscode.window.createTerminal("Playwright Install Terminal");
+            terminal.show(false);
+            terminal.sendText("npx playwright@1.38.0 install", true);
+            vscode.window.showInformationMessage("Please restart VS Code after installing Playwright browser");
+        } else {
+            vscode.window.showErrorMessage("Deactivating WebLab for VS Code");
+        }
     }
 
     registerCommands(context: vscode.ExtensionContext) {
@@ -37,7 +67,7 @@ export class BrowserProvider {
             "weblab-vscode.openBrowser",
             async () => {
                 const username = await this.getUsername();
-                const page = await this.browserContext.newPage();
+                const page = await this.getBrowserContext().newPage();
                 await page.goto(`https://weblab.tudelft.nl/profile/${username}`);
             }
         );
@@ -61,7 +91,7 @@ export class BrowserProvider {
             });
             this.context.secrets.store("password", password ?? "");
         }
-        const page = await this.browserContext.newPage();
+        const page = await this.getBrowserContext().newPage();
         await page.goto("https://weblab.tudelft.nl/samlsignin");
         await page.fill('input[name="username"]', username);
         await page.fill('input[name="password"]', password);
@@ -102,6 +132,10 @@ export class BrowserProvider {
     }
 
     getBrowserContext(): BrowserContext {
+        if (!this.browserContext) {
+            vscode.window.showErrorMessage("No browser initialized! Is WebLab enabled?");
+            throw new Error("No browser context initialized!");
+        }
         return this.browserContext;
     }